pyproject.toml
[project]
name = "arxplain"
version = "0.1.0"
description = "Generate easy explanation of any research paper"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "openai>=2.21.0",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
]
